clk: fix data type for aarch64
'offset' is used as a fully 32-bit/64-bit address at clk driver.
but it is declared as 32-bit and so this casues kernel panic at
suspend.
This patch changes u32 to 'unsigned long' for avoid this aarch64
issue. Also 'void __iomem *' causes build error at add operation
between 'void __iomem *'.

+++ drivers/clk/samsung/composite.h
@@ -22,8 +22,8 @@
  * @value: the value to be register at offset.
  */
 struct samsung_clk_reg_dump {
-	u32	offset;
-	u32	value;
+	unsigned long	offset;
+	u32		value;
 };
 
 /*
